+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 
 - Add `STAC::CommonMetadata` and make `Asset` and `Properties` include it.
+- Make `http_client` instance variable of `STACObject` instead of class instance variable of `ObjectResolver`.
 
 ## [0.1.0] - 2022-10-09